backend/requirements/base.txt
# Core
Django>=5.0,<5.1
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
python-decouple>=3.8
psycopg2-binary>=2.9.9

# API Documentation
drf-spectacular>=0.27.1

# Real-time
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0

# Tasks
celery>=5.3.6
redis>=5.0.1
django-celery-results>=2.5.1
django-celery-beat>=2.5.0

# Firebase
firebase-admin>=6.5.0

# Geolocation
django-filter>=24.2

# Storage & Quality
django-storages[s3]>=1.14.2
boto3>=1.34.0
Pillow>=10.2.0
whitenoise>=6.6.0
sentry-sdk>=1.40.6
django-lifecycle>=1.1.2
